requirements.txt
Django~=3.2.10
django-bootstrap-datepicker-plus==4.0.0
django-bootstrap-v5==1.0.11
django-import-export==2.4.0
importlib-metadata==5.0.0
mysql-connector==2.2.9
mysqlclient==2.0.2
